GEODE-5530 Add geode-web-test subproject for test-structure reuse
