+++ geode-web/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/FunctionCommandsOverHttpDUnitTest.java
@@ -21,7 +21,7 @@ import org.apache.geode.test.junit.categories.GfshTest;
 import org.apache.geode.test.junit.rules.GfshCommandRule;
 
 @Category({GfshTest.class})
-public class FunctionCommandsOverHttpDUnitTest extends FunctionCommandsDUnitTest {
+public class FunctionCommandsOverHttpDUnitTest extends FunctionCommandsDUnitTestBase {
   @Override
   public void connectGfsh() throws Exception {
     gfsh.connectAndVerify(getLocator().getHttpPort(), GfshCommandRule.PortType.http);
