+++ geode-web/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/GcCommandOverHttpDUnitTest.java
@@ -16,7 +16,7 @@ package org.apache.geode.management.internal.cli.commands;
 
 import org.apache.geode.test.junit.rules.GfshCommandRule;
 
-public class GcCommandOverHttpDUnitTest extends GcCommandDUnitTest {
+public class GcCommandOverHttpDUnitTest extends GcCommandDUnitTestBase {
   @Override
   public void before() throws Exception {
     gfsh.connectAndVerify(locator.getHttpPort(), GfshCommandRule.PortType.http);
