+++ geode-web/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/ListIndexCommandOverHttpDUnitTest.java
@@ -22,7 +22,7 @@ import org.apache.geode.test.junit.categories.GfshTest;
 import org.apache.geode.test.junit.rules.GfshCommandRule;
 
 @Category({GfshTest.class})
-public class ListIndexCommandOverHttpDUnitTest extends ListIndexCommandDUnitTest {
+public class ListIndexCommandOverHttpDUnitTest extends ListIndexCommandDUnitTestBase {
   @Override
   public void connectGfsh(MemberVM vm) throws Exception {
     gfsh.connectAndVerify(vm.getHttpPort(), GfshCommandRule.PortType.http);
