+++ geode-web/src/distributedTest/java/org/apache/geode/management/internal/cli/commands/RegionMembershipMBeanOverHttpDUnitTest.java
@@ -16,7 +16,7 @@ package org.apache.geode.management.internal.cli.commands;
 
 import org.apache.geode.test.junit.rules.GfshCommandRule;
 
-public class RegionMembershipMBeanOverHttpDUnitTest extends RegionMembershipMBeanDUnitTest {
+public class RegionMembershipMBeanOverHttpDUnitTest extends RegionMembershipMBeanDUnitTestBase {
   @Override
   public void connectToLocator() throws Exception {
     gfsh.connectAndVerify(locator.getHttpPort(), GfshCommandRule.PortType.http);
