+++ settings.gradle
@@ -23,6 +23,7 @@ include 'geode-junit'
 include 'geode-dunit'
 include 'geode-core'
 include 'geode-web'
+include 'geode-web/geode-web-test'
 include 'geode-web-api'
 include 'geode-pulse'
 include 'geode-assembly'
